requirements.txt
absl-py==2.4.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
cffi==2.0.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
flatbuffers==25.12.19
fonttools==4.61.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
mediapipe==0.10.9
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
packaging==26.0
parso==0.8.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pynput==1.7.7
pyobjc-core==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-Quartz==12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
scikit-learn==1.5.2
scipy==1.17.0
six==1.17.0
sounddevice==0.5.5
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.5.0
